+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.7.3] - 2019-09-10
+- Removed Dependency Validation test to prevent asking for major version changes when adding or removing dependencies
+- Fixed issues with scope of references used in APIValidation Assembly
+
 ## [0.7.2] - 2019-08-27
 - Add support for 2018.3 (without the package manager UI integration).